+++ package.json
@@ -4,12 +4,14 @@
   "private": false,
   "dependencies": {
     "axios": "^0.21.1",
+    "formik": "^2.2.6",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "styled-components": "^5.2.1",
-    "web-vitals": "^1.1.0"
+    "web-vitals": "^1.1.0",
+    "yup": "^0.32.9"
   },
   "scripts": {
     "format": "prettier --write \"**/*.{js,jsx,html,scss}\"",
