+++ src/pages/RegistrationForm/initialValues.js
@@ -1,0 +1,10 @@
+export const initialValues = {
+	name: '',
+	registerNumber: '',
+	email: '',
+	phoneNumber: '',
+	year: '',
+	gender: '',
+	dept: '',
+	avatar: ''
+};
