+++ src/pages/index.js
@@ -1,3 +1,3 @@
 export { default as Home } from './home';
 export { default as About } from './about';
-export { default as Register } from './register';
+export { default as Register } from './RegistrationForm/register';
